created a common navbar.js
